lint changes

+++ bindings/netstandard/ElectionGuard/ElectionGuard.ElectionSetup/Extensions/EnumerableExtensions.cs
@@ -2,8 +2,8 @@ namespace ElectionGuard.ElectionSetup.Extensions;
 
 public static class EnumerableExtensions
 {
-    public static IEnumerable<(T item, int index)> WithIndex<T>(this IEnumerable<T> source)
+    public static IEnumerable<(T item, ulong index)> WithIndex<T>(this IEnumerable<T> source)
     {
-        return source.Select((item, index) => (item, index));
+        return source.Select((item, index) => (item, (ulong)index));
     }
 }
